+++ include/interactive_mid_protocols/SigmaProtocolDamgardJurikEncryptedZero.hpp
@@ -232,7 +232,7 @@ private:
 	*/
 	bool checkChallengeLength(vector<byte> challenge) {
 		//If the challenge's length is equal to t, return true. else, return false.
-		return (challenge.size() == (t / 8) ? true : false);
+		return ((int)challenge.size() == (t / 8) ? true : false);
 	}
 
 public:
