+++ src/infra/Common.cpp
@@ -163,7 +163,7 @@ biginteger decodeBigInteger(byte* input, size_t length)
 
 biginteger convert_hex_to_biginteger(const string & input) {
 	string s = "0x" + input;
-	int index = s.find(" ");
+	size_t index = s.find(" ");
 	while (index != string::npos) {
 		s = s.substr(0, index) + s.substr(index + 1);
 		index = s.find(" ");
